Added locks when working with the variable conn #PSBM-152468
The conn variable is responsible for connection when working
with dbus from libvcmmd. When working with conn from multiple
threads, it is possible to race for data when the conn
variable is assigned the value NULL. This causes the assert
to be triggered in dbus and the subsequent drop in virtqemud.
We add locks when working with this variable from multiple threads.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>